# docs/openapi.json
{
  "info": {
    "description": "Personal finance tracker API.\n\nRequests that change data (POST, PUT, PATCH, DELETE) must send any body as `application/json` (415 otherwise). A browser may send them only from the same origin or an origin listed in `CORS_ORIGINS` (403 otherwise).",
    "title": "Vipu API",
    "version": "0.1.0"
  },
...
  "servers": [
    {
      "name": "Local",
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/income/{income_id}": {
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "income_id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "tags": [
          "Income"
        ],
        "summary": "Update an existing income item."
      },
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "income_id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "tags": [
          "Income"
        ],
        "summary": "Delete an income item."
      }
    },
...
    "/api/income/{income_id}/occurrence": {
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "income_id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "tags": [
          "Income"
        ],
        "summary": "Mark one occurrence of an income item as received, or as still expected.",
        "description": "Body: occurrence_date (YYYY-MM-DD) and settled (bool). Covers pay that\narrives early (a banking holiday moving payday forward) and pay that hasn't\nlanded on time, without moving the schedule: only that occurrence changes.\n\nThe date must be either the item's next occurrence or the most recent one\nthat has come due in the current pay period."
      }
    },
...
  },
  "openapi": "3.1.0",
  "components": {
    "schemas": {
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      }
    }
  }
}